fix substitution bug
I must have forgot to wrap the binder in a Var somewhere in the
substitution/reconstruction code. Problem solved by passing a Var
to body, instead of self, so that there is no chance of direct
nesting that will create a cycle in the graph, making its inter-
pretation as tree infinite.